Revised the way I was parsing string content.
I was trying to be clever and handle parsing sting content as a non-terminal on
its own. The problem is that the way I was doing it, I was actually trying to
lex it into a whole token (without even the delimiters), and that yields a
fundamental ambiguity when you want to tokenize simple stuff like `A::=c`, which
will end up being tokenized into one big blob under the assumption that it is
text content.

After reading https://github.com/lalrpop/lalrpop/issues/27
I realized that the right way to do this is to treat the string literal
as one big token, *including* the delimiters.

(I also decided to rip out the way I was handling escape sequences, though I
think I can actually put them back in easily now that I have thought about it
for a moment.)