get the current set of tests (up through imperative fixed-width integer) to pass.
The biggest shift this required was moving from the fiction that parsing is
deterministic (and returns `Option<Env>`) to making it reflect its
non-deterministic nature, and so have it return `Iterator<Item=Env>`. But then
its not just an `Iterator` (since that's a trait), but rather a
`Box<dyn Iterator<Item=Env>>`.

That change had implications for things like how `Concat(l,r)` is interpreted.
When the result is just an `Option`, you can drive forward from left-to-right,
it is easy. But if the left side returns an iterator, you need to potentially
backtrack. Worse still, the environment you get back from parsing `l` needs to
be fed into the parse attempt for `r`; that means that the typical methods for
iterator composition *do not work*.

I made my own `Cross` iterator to deal with this, which is like
`std::iter::Zip`, except that unlike `Zip`, the right-hand iterator used for
constructing the tuples is *generated via* the result of running the left-hand
iterator. But then I ended up *not* using that `Cross` iterator when
interpreting `Kleene`; instead I made a special helper method, `split_matches`,
that I think would work quite well for `Concat` as well. If I go down that road,
I should be able to get rid of `Cross`.

Changing the return type of a recursive algorithm like this required many changes
throughout the code; e.g. it being a `dyn Trait` meant that I had to decide
whether it would obey a 'static bound, or if it would admit a broader set of
lifetimes. (For now, it is bounded by the lifetime attached to the `&self`
parameter of the `Grammar`.

It also meant I had to change the definiton of `Blackbox` (which is still
untested); it now carries a `fn()` pointer rather than a `Box<dyn Fn>` (mainly
because I had to be able to clone black boxes). This is almost certainly fine;
the purpose of blackbox is to link to predefied parsing routines; it seems very
unlikely that those would be closures in practice. In any case, I'll deal with
this later; another obvious way that I'm now only seeing in hindsight would be
to carry a `&dyn Fn` instead of a `Box<dyn Fn>`.

Overall, I'm not committed to any of the decisions made here. There's lot of
places where I'm cloning, and a lot of places where I'm wrestling with lifetimes
that seem like just a design mistake.

I also made a number of `std::fmt::Display` impls because a needed a more
succinct output format than the standard `Debug` format when trying to
understand what the algorithm was doing in practice.