Checkpoint status. Started sketching transducer structure itself.
Realized that the (non-deterministic) spec for stack-eval of transducers itself
also belongs in the speciciation rules part of the code. Which then led me to
refactor that code slightly, since I do not expect much overlap at the moment
between what the grammar spec code looks like and what the transducer spec code
will look like when I fill it in.