Added method for following a call-transition on a tranducer configuration.
This was specifically driven by example of figure 2(b) in the paper, and thus I
deliberately am allowing for the input to be omitted.

I am also choosing to nonetheless parameterize over the choice of variable; this
may be ill-advised because the formalism fixes it at y_0, and most reasonable
systems will derive it from the "callee", not the caller (if they allow it to
vary at all, that is).