More of the trace of fig 2(b) on "1", now including a return transition, and a helper function to ease constructing the expected configrations from those transitions.
However, the process of making this test really made it clear that there is a
gap between the idealized tool that is presented in the initial examples
versus the diagrams in Figure 2 and versus the concrete semantics.

In particular, Fig 2(b) doesn't show how variable binding should be handled in a
`call` transition. The formal semantics omits formal parameters by always using
`y_0` on every call; but the latter *cannot* be the semantics intended in
fig~2(b), because `y_0` isn't used in fig~2(b), only `n`.

At this point I think the intention of the paper is that in figure 2(b), one can
tell from the accompanying grammar that the *intent* is to bind `n`. (And that
is not part of the formal semantics because choosing a formal parameter name is
entirely "uninteresting" if one does not have nesting of binding forms. But,
having said that, I also think that I want to "fix" things now to allow me to
select names for the formal parameter, if any, of each non-terminal.