Added support for customizing the formal parameter name in a tranducer.
There are two ways on can imagine doing this: extend the call transitions so
that they can select the parameter name for the input they feed in, or extend
the states themselves to choose what name to use if they are the target of a
call transition.

I went with the latter option. So conceptually, figure 2(b) can be thought of as
being extended so that the state labelled "1" has the additional variable name
"n" attached to it, and that's used for the input that is fed in via the
"call(n-1)" transition from state 3.