Finished the test of encoding of fig~2(b) on input "1".
It was a good thing I started trying to observe the end parse tree; I was
briefly going to stop doing an earlier "accept" state, but it was an accept of
the non-terminal `int(0)` (i.e. a zero-length string), which pops *out* to the
caller configuration that then accepts the `int(1)` that is the appropriate
non-terminal here.

Fixing this did also force me to fix an oversight in my earlier addition of the
formal parameter support: the whole thing about needing to check that the input
value matches what the environment has for `y_0` was causing the rule to not
fire, and I assume that was in part because I am no longer binding the input to
`y_0` in the first place.

So I had to generalize that rule accordingly; luckily the presence of the state
for the call entry in the configuration state makes this easy, thought it deos
make me wonder about the other part of the paper that pretty explicitly claimed
that that the state for the call entry is effectively a bit of ghost state
solely here to make another proof easier to do... that's no longer true for me,
since I'm extracting the formal parameter name *from* that call context.